requirements.txt
langgraph>=0.2.55
langchain-community>=0.3.9
langchain-openai>=0.3.7
langchain-anthropic>=0.3.9
langchain-deepseek>=0.1.2
openai>=1.61.0
duckduckgo_search>=7.5.0
langchain-groq>=0.2.4
exa-py>=1.8.9
arxiv>=2.1.3
pymupdf>=1.25.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
dashscope>=1.22.2
chainlit==2.3.0
tavily-python==0.5.1